Split docker cache from base tarball caching.
For non arm builds, the base tarballs should be invalidated when
the changelog changes.

For arm builds, the cache should be invalidated when the mini-root
changes.

We still need to follow up the mini-root conversion with a slackpkg
update for the arm builds, but we first need to find the solution to
this faccessat2 syscall problem in the arm 15.0 and current images.