Temporary workaround for arm faccessat2 issue.
The arm image does not work when run on amd64 (which is what the github
build is). faccessat2 raises, but does not get properly handled in a
fallback, leading to things not really working.

For example:

```sh
$ test -x /bin/sh
$ echo $?
1
```

Slackware and Slackware64 compile glibc with `--enable-kernel=2.6.32`
while slackware arm uses `--enable-kernel=5.10`.

The reason for this is:

https://github.com/bminor/glibc/blob/b92a49359f33a461db080a33940d73f47c756126/sysdeps/unix/sysv/linux/kernel-features.h#L217

which affects this implementation of this method:

https://github.com/bminor/glibc/blob/6f043e0ee7e477f50a44024ed0cb579d5e3f511d/sysdeps/unix/sysv/linux/faccessat.c#L26-L76

I will see what Stuart thinks about this, but for now, I am pulling the
glibc from ubuntu so that the arm image actually works.

I will update the Dockerfile to rebuild glibc in a followup commit.