Drop linux/386 entrypoint config.
This is non-standard - other distros do not do this - uname will return
x86_64 on an x86_64 host for a linux/386 image for ubuntu, debian,
alpine and more, so let's match this.

If you want the previous behaviour, it is simple to add an entrypoint
flag to your docker run command.