Prefer shell variables for environment input in GHA. (#16038)
If a value is already set in the env, we can access it natively in shell
instead of using `${{ env.KEY }}`. The latter is done after parsing the
yaml and before it is passed to shell script and requires careful
escaping.

None of these are actually concerning in our case since they are not
user input, but zizmor checks flag them and we'd like to enable more
zizmor checks, but these existing violations are currently blocking
that.

These should all be straightforward fixes where we either were in
javascript or shell and were allowing values to be interpolated in via
`${{ env.KEY }}`.